feat(key-auth) update key-auth's param duplication detection to new encoding
This is both a feat and a fix.

The duplication code for JSON-encoded requests remains unchanged: if `apikey` is a table instead of a string, the "Duplicate key" error is thrown.

The form-url-encoded is the one which changes. With the recent changes in params encoding & decoding, when this body is encountered:

```
{apikey = { "kong", "kong" } }
```

It was encoded like this (with no_array_indexes):
```
apikey=kong&apikey=kong
```
And like this (with array indexes):
```
apikey[1]=kong&apikey[2]=kong
```

The "fix" part of this PR is that that second case (with array indexes) is now detected and the proper "Duplicated error" is returned.

The "feat" part of this PR is that now, the no_array_indexes option encodes the urls with brackets:
```
apikey[]=kong&apikey[]=kong
```

Which is now also properly detected on this change.